fix: don't spawn powerups immediately
Closes https://github.com/ec-/baseq3a/issues/8.

The bug has been introduced in
2c65e577e393a8151c74a84130300ff13a9617cc.
This commit basically brings back the code that was removed
as part of that one.

The intent of that commit was to spawn the powerups immediately
for warmup, but then spawn them as usual when the warmup ends,
in `G_WarmupEnd`.
However, it's not always that we get to execute `G_WarmupEnd`.
Namely, it doesn't get executed when `g_warmup 0`
or in single player matches.

The downside is, of course, that we no longer spawn the powerups
for warmup immediately.

So let's simply revert that change until a better solution is found.